+++ packages/router_history/src/index.ts
@@ -1,6 +1,5 @@
 import {
     BasePluginType,
-    BrowserStackTypes,
     EventTypes,
     ReportDataType,
     RouteDataMsgType,