+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @tc-track/browser
 
+## 1.0.0
+
+### Major Changes
+
+-   项目初始化
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @tc-track/types@1.0.0
+    -   @tc-track/utils@1.0.0
+    -   @tc-track/core@1.0.0
+
 ## 1.1.10
 
 ### Patch Changes